backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Banco de dados
sqlalchemy==2.0.23
alembic==1.12.1

# IA e NLP
google-generativeai==0.3.1
edge-tts==6.1.9
pyttsx3==2.90

# Audio
playsound==1.3.0

# TTS - Síntese de Voz
gtts==2.5.4  # Google TTS para PT-BR (recomendado)
TTS>=0.22.0  # Coqui TTS (VITS) - Melhor qualidade PT-BR

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0

# CORS
python-multipart==0.0.6

# Logging e monitoramento
python-json-logger==2.0.7

# Cache (opcional)
redis==5.0.1

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
